Predict effector

predict_effector is used to predict effector protein given amino acid protein sequences.

predict_effector(input, taxon)

# S3 method for character
predict_effector(input, taxon)

# S3 method for data.frame
predict_effector(input, taxon)

# S3 method for AAStringSet
predict_effector(input, taxon)

# S3 method for AAString
predict_effector(input, taxon)

# S3 method for default
predict_effector(input, taxon)

Arguments

input
Input data that contains amino acid sequence(s). It can be in fasta format, strings, AAString, AAStringset, and dataframe.

taxon
Taxon group of input data. Available taxons are bacteria, fungi, and oomycete.

Value

predict_effector returns an object of class "tbl_deepredeff" or for multiple responses of class c("tbl_deepredeff", "data.frame").

An object of class "tbl_deepredeff" is a data frame containing at least the following components:

sequence

the sequence(s) from the input data.

s_score

score obtained from sigmoid function showing how likely the sequences to be an effector.

prediction

class prediction for each sequence, obtained from s_score. If the value of s_score >= 0.5, it will be classified as an effector. Otherwise, it will be classified as a non-effector.

Examples

if (FALSE) {
# FASTA input
input_fasta <- system.file("extdata/example/fungi_sample.fasta", package = "deepredeff")

pred_result <- deepredeff::predict_effector(
  input = input_fasta,
  taxon = "fungi"
)
}